requirements.txt
Django==2.1.2
django-heroku
gunicorn
python-decouple
whitenoise
Pillow
dj_database_url